GUI responsive when device disconnected
But will not respond until device has been connected at least once!  On
a side note I found out how silly the saveState/loadState method in
genericUsbDriver was.  It was always a better idea to just "poke" the
signals to ensure that it is reset in the same way that would happen
from GUI interactions.